put the annotate widget in a dedicated file and add more help-texts and
